data-plane/python-bindings/examples/pyproject.toml
# Copyright AGNTCY Contributors (https://github.com/agntcy)
# SPDX-License-Identifier: Apache-2.0

[project]
name = "slim-bindings-examples"
version = "0.1.0"
description = "Example apps using SLIM Rust bindings for Python"
requires-python = ">=3.12"
license = "Apache-2.0"
dependencies = [
    "slim-bindings>=0.3.6",
    "pyjwt>=2.10.1",
]

[tool.uv.sources]
slim-bindings = { workspace = true }

[project.scripts]
examples = "slim_bindings_examples:main"
slim = "slim_bindings_examples:slim_main"
fire-and-forget = "slim_bindings_examples:fire_and_forget_main"
request-reply = "slim_bindings_examples:request_reply_main"
streaming = "slim_bindings_examples:streaming_main"
pubsub = "slim_bindings_examples:pubsub_main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/slim_bindings_examples"]
